requirements.txt
streamlit==1.44.1
pdf2image==1.17.0
pytesseract==0.3.13
langchain-ollama=0.3.1
langchain==0.3.23
langchain_core==0.3.51
pillow==11.2.1